Updated Circle Movement
I have an interesting movement with the circles right now, right now the circles look like they are being visually swallowed by a circle at mouseX and mouseY, this wasn't my original intention originally I wanted the movement to mirror the movement of a flock of swallows and keep a certain amount of spacing but now it's kind of cool because the swallows are being swallowed. I don't know if I want to keep the original shape as this pyramid and scrap the bird swallow idea but I kind of like it when there is a large number of circles, when there's like 900 circles the effect turns into a kind of waving ribbon effect.